% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{importDimLabel}
\alias{importDimLabel}
\title{Import Eurostat label (description) of a given dimension code}
\usage{
importDimLabel(EurostatDimCode)
}
\arguments{
\item{EurostatDimCode}{A string -- the code name of the Eurostat dimension, e.g. \code{"geo"} or \code{"nace_r2"}
or \code{"indic_sb"}, etc.}
}
\value{
A character vector of length 1: the label/description of \code{EurostatDimCode}.
}
\description{

}
