% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_search.R
\name{browseDataList}
\alias{browseDataList}
\title{Search Eurostat datasets and see the result as a table in a browser}
\usage{
browseDataList(...)
}
\arguments{
\item{...}{An expression (symbolic, unquoted) to be passted to \code{\link{subset}}.
The column names of the table of datasets can be used -- those with spaces should be
backtick (`) quoted. See the examples below.}
}
\value{
\itemize{
  \item Side effect -- a table opened in a browser via \code{\link[utils]{browseURL}}.
  \item Value -- a list with:
  \itemize{
    \item df -- a data.frame, imported via \code{\link[eurodata]{importDataList}} and
    filtered based on the conditions specified in \code{...}.
    \item html -- a string, with the HTML code that generated the table in a browser.
  }
}
}
\description{
Search Eurostat datasets and see the result as a table in a browser
}
\examples{
\dontrun{
browseDataList(grepl('servic',`Dataset name`))
browseDataList(grepl('bop',Code) & !grepl('its',Code))
}
}

