% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_import.R
\name{importData}
\alias{importData}
\title{Download and import a Eurostat dataset}
\usage{
importData(EurostatDatasetCode)
}
\arguments{
\item{EurostatDatasetCode}{A string with Eurostat dataset code name, e.g. \code{nama_10_gdp} or \code{bop_its6_det}.
See e.g.:
\url{http://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?sort=1&file=table_of_contents_en.pdf}}
}
\value{
A Eurostat dataset as a `flat' data.frame.
A `flat' dataset has all numeric values in one column, with each row representing one of the available combinations
of all dimensions (e.g. if dimensions are: countries, years, sectors, and indicators, there can be a row for value
added in retail in Germany in 2013).
}
\description{
Download and import a Eurostat dataset
}
