% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_import.R
\name{importDataList}
\alias{importDataList}
\title{Import and reshape Eurostat inventory of datasets}
\usage{
importDataList()
}
\value{
The imported data.frame reflects the hierarchical
structure of datasets (see the columns \code{Data subgroup, level 0},
\code{Data subgroup, level 1}, \code{Data subgroup, level 2}, etc.).
It is tagged with S3 class \code{EurostatDataList}.
}
\description{
Import and reshape Eurostat inventory of datasets
}

